pyproject.toml
[project]
name = "spotify-mcp-server"
version = "0.1.0"
description = "Spotify MCP server to connect to LLM agent"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask>=3.1.1",
    "httpx>=0.28.1",
    "mcp[cli]>=1.9.4",
    "python-dotenv>=1.1.0",
    "spotipy>=2.25.1",
]
